summarize/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
diskcache==5.6.1
filelock==3.12.2
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
llama-cpp-python==0.1.77
numpy==1.25.1
packaging==23.1
PyYAML==6.0.1
requests==2.31.0
tqdm==4.65.0
typing_extensions==4.7.1
urllib3==2.0.4
